moved client contents into __init__.py files to avoid importing via the client module
