+++ Makefile
@@ -26,18 +26,16 @@ install:
 ##
 test:
 
-##		Upload to testpypi
 ##
-test_upload:
-	twine upload -r testpypi dist/*
-
-##		Upload to PyPI
 ##
-live_upload:
-	twine upload dist/*
+doc:
+	cd docs && make html
 
 ##		Clean the workspace
 ##
 clean:
-	rm -rf .pytest_cache build/ dist/ python_kraken_sdk.egg-info
-	rm -f .coverage kraken/_version.py
+	rm -rf .pytest_cache build/ dist/ python_kraken_sdk.egg-info docs/_build
+	rm -f .coverage kraken/_version.py *.log *.csv *.zip tests/*.zip tests/.csv
+	find tests -name "__pycache__" | xargs rm -rf
+	find kraken -name "__pycache__" | xargs rm -rf
+	find examples -name "__pycache__" | xargs rm -rf
