+++ tests/__init__.py
@@ -2,4 +2,5 @@
 # -*- coding: utf-8 -*-
 # Copyright (C) 2023 Benjamin Thomas Schwertfeger
 # Github: https://github.com/btschwertfeger
+#
 # This file is required for the CI/CD codecov workflow.
